requirements.txt
# Core
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-chroma==0.1.2
chromadb==0.5.4

# Embeddings (Gemini)
langchain-google-genai==2.0.0

# LLMs (pick at least one)
google-genai==0.6.0
openai==1.51.2
langchain-openai==0.1.22

# Reranker
sentence-transformers==3.0.1
transformers==4.43.3
accelerate==0.33.0

# PDF & parsing
pypdf==4.3.1
unstructured==0.16.5

# UI
chainlit==2.0.1

# Eval
ragas==0.1.21
scikit-learn==1.5.1
pandas==2.2.2

# Crawling
beautifulsoup4==4.12.3
httpx==0.27.0

# Text splitters
langchain-text-splitters==0.2.2

# MCP (optional; install separately if needed)
python-dotenv==1.0.1

pydantic-settings>=2.0.0